loading last project on startup

+++ SportsAnalysis/models/ProjectCode.swift
@@ -11,10 +11,9 @@ struct ProjectCode : Codable, Hashable {
     var name: String
     var color: Color
     
-    // TODO change type at some point
     var shortcut: String
-    var leadingTime: Double = 10
-    var trailingTime: Double = 10
+    var leadingTime: Float = 10
+    var trailingTime: Float = 10
     
     init(name: String, color: Color = .white, shortcut: String) {
         self.name = name
